Removed type lr_off_t

+++ librepo/curl.h
@@ -44,7 +44,7 @@ extern "C" {
 #define lr_curl_single_download(HANDLE, URL, FD) \
             lr_curl_single_download_resume((HANDLE), (URL), (FD), 0)
 
-int lr_curl_single_download_resume(lr_Handle handle, const char *url, int fd, long offset);
+int lr_curl_single_download_resume(lr_Handle handle, const char *url, int fd, long long offset);
 
 #define lr_curl_single_mirrored_download(HANDLE, FILENAME, FD, CHKSUM_T, CHKSUM)\
             lr_curl_single_mirrored_download_resume((HANDLE), \
@@ -59,7 +59,7 @@ int lr_curl_single_mirrored_download_resume(lr_Handle handle,
                                             int fd,
                                             lr_ChecksumType checksum_type,
                                             const char *checksum,
-                                            long offset);
+                                            long long offset);
 
 int lr_curl_multi_download(lr_Handle handle, lr_CurlTargetList targets);
